Create The Double Color Ball
I designed a program based on conditional probability to buy the lottery-The Double Color Ball. The first prize principle of the bicolor ball is that choosing 6 numbers from 33 red balls, and 1 number from 16 blue balls, and if those numbers are matched with the official ones, then the buyer wins the first prize. To raise the winning probability, I designed the conditional probability-based algorithm in this way: taking the red ball for example, I first defined the array A constituted by numbers from 1 to 33, then found out the numbers not winning prizes in recent 10 times and added them into array A. Now, there were about 40 red balls in array A. Afterward, based on the theory of conditional probability, I deleted numbers that had won prize last time. Finally, by using the random function, I selected 6 numbers from array A, if repeated numbers were existing, I would delete repeated numbers and record times as a, and then generated randomly continuously until obtained 6 different numbers. And the operation of the blue ball was in the same way.